cli: list-plugins cmd migrated to docopts.
