+++ snapcraft/cmds.py
@@ -118,16 +118,6 @@ def assemble(args):
         sys.exit(ret)
 
 
-def list_plugins(args=None):
-    import pkgutil
-    import snapcraft.plugins
-
-    for importer, modname, is_package in pkgutil.iter_modules(
-            snapcraft.plugins.__path__):
-        if not is_package:
-            print(modname.replace('_', '-'))
-
-
 def clean(args):
     config = _load_config()
